Handle missing mailing list archive
When there are no gzipped mailing list archives, mail_summary()
passes curl's 404 HTML contents to zcat which promptly shows
the user "gzip: stdin: not in gzip format".

Allow for the possibility that this can happen briefly at the
beginning of a new month (no mail messages to the mailing list),
and don't bother showing the zcat error to the user: it doesn't
help them understand what's going on, and there's nothing they
can do.